Add eslint config to be used in most JavaScript projects
Why:

* We want our JS style to be uniform across projects.

* This configuration makes two assumptions about the project: (a) it is
using babel, (b) it is using react. It may not be the case for every
project, but because this is the direction we are moving towards it is
justified to have it on the configuration.

* There is a third assumption: it is a front-end project. When working with node
you don't want to transpile the code. There are already some es6
features available, but this configuration will not work. For that I'll
later submit a different configuration.

This change addresses the need by:

* Adding a `.eslintrc` file.